+++ extensions/cloud/portability-cloud-google/build.gradle
@@ -15,6 +15,6 @@
  */
 
 dependencies {
-    compile project(':portability-spi-data-transfer')
     compile project(':portability-spi-cloud')
+    compile project(':portability-spi-worker')
 }
