--- portability-spi-data-transfer/src/main/java/org/dataportabilityproject/spi/datatransfer/Exporter.java
@@ -1,11 +1,0 @@
-package org.dataportabilityproject.spi.datatransfer;
-
-import org.dataportabilityproject.datatransfer.types.models.DataModel;
-
-/**
- * Placeholder.
- */
-public interface Exporter<T extends DataModel> {
-
-    T export(Object continuationInformation); // REVIEW: The original throws IOException. Continue to use checked exceptions or use unchecked?
-}
