+++ portability-spi-worker/src/main/java/org/dataportabilityproject/spi/worker/Exporter.java
@@ -13,8 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.dataportabilityproject.spi.worker;
 
-dependencies {
-    compile project(':portability-spi-cloud')
-    compile project(':portability-spi-worker')
+import org.dataportabilityproject.datatransfer.types.models.DataModel;
+
+/**
+ * Placeholder.
+ */
+public interface Exporter<T extends DataModel> {
+
+    T export(Object continuationInformation); // REVIEW: The original throws IOException. Continue to use checked exceptions or use unchecked?
 }
