+++ portability-spi-worker/src/main/java/org/dataportabilityproject/spi/worker/Importer.java
@@ -13,8 +13,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- * Defines the data portability SPI. Extensions use this module to implement custom functionality for the system.
- */
-package org.dataportabilityproject.spi.datatransfer;
+package org.dataportabilityproject.spi.worker;
 
+import org.dataportabilityproject.datatransfer.types.models.DataModel;
+
+/**
+ * Imports data into a destination service.
+ */
+public interface Importer<T extends DataModel> {
+
+    /**
+     * Returns the service provider id of the importer.
+     */
+    String getServiceProviderId();
+
+    /**
+     * Returns supported content types.
+     */
+    String[] getContentTypes();
+
+    /**
+     * Imports data.
+     *
+     * @param data the data
+     */
+    void importItem(T data); // REVIEW: The original throws IOException. Continue to use checked exceptions or use unchecked?
+
+}
